check for empty updates
We basically count archs up and down, and if the overall count is != 0,
then we have to do something (update or delete). We might have a manifest list
without any archs left (count <0). In this case the manifest list should
be deleted. For now we don't delete, but skip the generation (which
 would produce an invalid "docker" command line anyways). If all archs
get deleted, one should also delete the manifest list. Up to the admin.