Rewrite to use crane
This commit switches the implementation from github.com/genuinetools/reg
to using github.com/google/go-containerregistry ("crane").

The main issue that needed addressing was that reg was no longer updated, and
caused some issues with some uploaded images breaking prestomanifesto. A switch
to crane should remedy this issue.

Since we now create the manifest lists explicitly, we can skip some checks and
simply try to recreate the manifests every time: since the resulting manifest
should have the same form every time, it will also have the same digest. The
push is then skipped, as the manifest already exists. This means we no longer
need to compare the existing manifests to the available arch-specific images.

We also make the actual merging into a manifest list a separate package,
as that may be useful for other packages, for example running the merge
step directly on push.